[#283] refactor: BadgeMatchedCountTemplate으로 이름 변경에 따른 변
