+++ src/main/java/goorm/eagle7/stelligence/domain/badge/template/impl/ContributeRejectedBadgeMatchedCountTemplate.java
@@ -1,5 +1,7 @@
 package goorm.eagle7.stelligence.domain.badge.template.impl;
 
+import static goorm.eagle7.stelligence.domain.badge.model.Badge.*;
+
 import java.util.HashMap;
 import java.util.Map;
 
@@ -7,37 +9,36 @@ import org.springframework.stereotype.Component;
 
 import goorm.eagle7.stelligence.domain.badge.model.Badge;
 import goorm.eagle7.stelligence.domain.badge.model.BadgeCategory;
-import goorm.eagle7.stelligence.domain.badge.template.BadgeAwardTemplate;
-import goorm.eagle7.stelligence.domain.document.content.DocumentContentRepository;
+import goorm.eagle7.stelligence.domain.badge.template.BadgeMatchedCountTemplate;
+import goorm.eagle7.stelligence.domain.contribute.ContributeRepository;
+import goorm.eagle7.stelligence.domain.contribute.model.ContributeStatus;
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class DocumentBadgeAwardTemplate extends BadgeAwardTemplate {
+public class ContributeRejectedBadgeMatchedCountTemplate extends
 
+	BadgeMatchedCountTemplate {
+
+	private final ContributeRepository contributeRepository;
 	private static final Map<Integer, Badge> requiredCounts = new HashMap<>();
-	private final DocumentContentRepository documentContentRepository;
 
 	@Override
 	public boolean supports(BadgeCategory category) {
-		return BadgeCategory.DOCUMENT.equals(category);
+		return BadgeCategory.CONTRIBUTE_REJECTED.equals(category);
 	}
 
 	@Override
 	protected long getCount(Long memberId) {
-		return documentContentRepository.countByAuthor_Id(memberId);
+		return contributeRepository.countByMemberIdAndStatus(memberId, ContributeStatus.REJECTED);
 	}
 
 	@Override
 	protected Map<Integer, Badge> getBadgeCriteria() {
 
 		if (requiredCounts.isEmpty()) {
-			requiredCounts.put(1, Badge.ASTRONAUT);
-			requiredCounts.put(5, Badge.MOON);
-			requiredCounts.put(10, Badge.MARS);
-			requiredCounts.put(20, Badge.URANUS);
+			requiredCounts.put(100, BLACKHOLE);
 		}
 		return requiredCounts;
 	}
-
 }
