+++ src/main/java/goorm/eagle7/stelligence/domain/badge/template/impl/DocumentBadgeMatchedCountTemplate.java
@@ -7,37 +7,37 @@ import org.springframework.stereotype.Component;
 
 import goorm.eagle7.stelligence.domain.badge.model.Badge;
 import goorm.eagle7.stelligence.domain.badge.model.BadgeCategory;
-import goorm.eagle7.stelligence.domain.badge.template.BadgeAwardTemplate;
-import goorm.eagle7.stelligence.domain.report.CommentReportRepository;
-import goorm.eagle7.stelligence.domain.report.DocumentReportRepository;
+import goorm.eagle7.stelligence.domain.badge.template.BadgeMatchedCountTemplate;
+import goorm.eagle7.stelligence.domain.document.content.DocumentContentRepository;
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class ReportBadgeAwardTemplate extends BadgeAwardTemplate {
+public class DocumentBadgeMatchedCountTemplate extends BadgeMatchedCountTemplate {
 
 	private static final Map<Integer, Badge> requiredCounts = new HashMap<>();
-	private final DocumentReportRepository documentReportRepository;
-	private final CommentReportRepository commentReportRepository;
+	private final DocumentContentRepository documentContentRepository;
 
 	@Override
 	public boolean supports(BadgeCategory category) {
-		return BadgeCategory.REPORT.equals(category);
+		return BadgeCategory.DOCUMENT.equals(category);
 	}
 
 	@Override
 	protected long getCount(Long memberId) {
-		long countDocument = documentReportRepository.countByReporterId(memberId);
-		long countComment = commentReportRepository.countByReporterId(memberId);
-		return countDocument + countComment;
+		return documentContentRepository.countByAuthor_Id(memberId);
 	}
 
 	@Override
 	protected Map<Integer, Badge> getBadgeCriteria() {
 
 		if (requiredCounts.isEmpty()) {
-			requiredCounts.put(10, Badge.GUARD);
+			requiredCounts.put(1, Badge.ASTRONAUT);
+			requiredCounts.put(5, Badge.MOON);
+			requiredCounts.put(10, Badge.MARS);
+			requiredCounts.put(20, Badge.URANUS);
 		}
 		return requiredCounts;
 	}
+
 }
